more updates to histogramming framework
big new feature is the support for jet cuts. Jet cuts are year- and
trigger-dependent. We're doing the trigger part using a filter generator in the
histos2 module, and passing that filter function to each config module in
analyze. Config modules now support **kwargs so we don't have to change old
modules if we want to send more info in analyze() in the future. The older
modules can just ignore the new args.

The year-dependent part is solved by calculating the year from the runId (in
data) or the muDstName (simulations) and setting the year as an attribute of the
event. Modules can do year-dependent cuts just by checking event.year.