new implementation of HistogramManager
It should be completely backwards-compatible w.r.t. read operations (writes are
of course accomplished using the new config framework). Advantages over the old
implementation include lazy loading of histograms (only the TKeys are read on
initialization now, so there should be no need to specify keys and triggers in
the ctor) and the ability to use the dotted attribute syntax to drill down into
the hierarchy (although integers can't be attributes, so specific triggers must
be retrieved using the [] syntax).